refactor: insertOrUpdateEmoji uploadEmojiCustom emoji-custom
